getting things lined up and arranged
