+++ src/dialogue/pages/page1/page1-index.js
@@ -1,0 +1,10 @@
+import Rx from 'rx'
+import view from './page1-view'
+
+const Page1 = () => {
+  return {
+    DOM: Rx.Observable.just(view())
+  }
+}
+
+export default Page1
