+++ src/dialogue/pages/page2/page2-index.js
@@ -1,0 +1,12 @@
+import Rx from 'rx'
+import view from './page2-view'
+
+const Page2 = () => {
+
+  const view$ = Rx.Observable.just(view());
+  return {
+    DOM: view$
+  }
+}
+
+export default Page2
