+++ src/utils/routes.js
@@ -1,7 +1,7 @@
 import {div, h1, h} from '@cycle/dom'
-import Home from '../dialogue/pages/home'
-import Page1 from '../dialogue/pages/page1'
-import Page2 from '../dialogue/pages/page2'
+import Home from '../dialogue/pages/home/home-index'
+import Page1 from '../dialogue/pages/page1/page1-index'
+import Page2 from '../dialogue/pages/page2/page2-index'
 
 const routes = {
   '/': div('.home', Home().DOM),
